refactor: created a DummyFactory to be responsible of creating dummies for the integration tests
