+++ web/src/test/kotlin/br/all/review/persistence/MongoSystematicStudyRepositoryTest.kt
@@ -1,9 +1,8 @@
 package br.all.review.persistence
 
 import br.all.infrastructure.review.MongoSystematicStudyRepository
-import br.all.infrastructure.review.SystematicStudyDocument
 import br.all.infrastructure.shared.toNullable
-import io.github.serpro69.kfaker.Faker
+import br.all.review.shared.DummyFactory
 import org.junit.jupiter.api.*
 import org.junit.jupiter.api.Assertions.*
 import org.springframework.beans.factory.annotation.Autowired
@@ -15,10 +14,11 @@ import java.util.*
 class MongoSystematicStudyRepositoryTest(
     @Autowired private val sut: MongoSystematicStudyRepository,
 ) {
-    private val faker = Faker()
+    private lateinit var dummyFactory: DummyFactory
 
     @BeforeEach
     fun setUp() {
+        dummyFactory = DummyFactory()
         sut.deleteAll()
     }
     @Nested
@@ -28,7 +28,7 @@ class MongoSystematicStudyRepositoryTest(
         @Test
         fun `Should save a new systematic study`() {
             val systematicStudyId = UUID.randomUUID()
-            val document = generateDocument(systematicStudyId)
+            val document = dummyFactory.createSystematicStudyDocument(systematicStudyId)
 
             sut.save(document)
             assertEquals(document, sut.findById(systematicStudyId).toNullable())
@@ -38,11 +38,11 @@ class MongoSystematicStudyRepositoryTest(
         fun `Should update an existent systematic study`() {
             val systematicStudyId = UUID.randomUUID()
             val ownerId = UUID.randomUUID()
-            val oldDocument = generateDocument(id = systematicStudyId, owner = ownerId)
+            val oldDocument = dummyFactory.createSystematicStudyDocument(id = systematicStudyId, owner = ownerId)
 
             sut.save(oldDocument)
 
-            val newDocument = generateDocument(
+            val newDocument = dummyFactory.createSystematicStudyDocument(
                 id = systematicStudyId,
                 owner = ownerId,
                 collaborators = mutableSetOf(UUID.randomUUID())
@@ -56,7 +56,7 @@ class MongoSystematicStudyRepositoryTest(
         @Test
         fun `Should find a existent systematic study`() {
             val systematicStudyId = UUID.randomUUID()
-            val document = generateDocument(systematicStudyId)
+            val document = dummyFactory.createSystematicStudyDocument(systematicStudyId)
 
             sut.save(document)
             assertNotEquals(null, sut.findById(systematicStudyId).toNullable())
@@ -64,9 +64,9 @@ class MongoSystematicStudyRepositoryTest(
 
         @Test
         fun `Should find all existent systematic study`() {
-            sut.save(generateDocument())
-            sut.save(generateDocument())
-            sut.save(generateDocument())
+            sut.save(dummyFactory.createSystematicStudyDocument())
+            sut.save(dummyFactory.createSystematicStudyDocument())
+            sut.save(dummyFactory.createSystematicStudyDocument())
 
             assertEquals(3, sut.findAll().size)
         }
@@ -74,7 +74,7 @@ class MongoSystematicStudyRepositoryTest(
         @Test
         fun `Should delete a systematic study`() {
             val systematicStudyId = UUID.randomUUID()
-            val document = generateDocument(systematicStudyId)
+            val document = dummyFactory.createSystematicStudyDocument(systematicStudyId)
 
             sut.save(document)
 
@@ -97,12 +97,4 @@ class MongoSystematicStudyRepositoryTest(
             assertEquals(0, sut.findAll().size)
         }
     }
-
-    private fun generateDocument(
-        id: UUID = UUID.randomUUID(),
-        title: String = faker.book.title(),
-        description: String = faker.lorem.words(),
-        owner: UUID = UUID.randomUUID(),
-        collaborators: MutableSet<UUID> = mutableSetOf(),
-    ) = SystematicStudyDocument(id, title, description, owner, collaborators.also { it.add(owner) }.toSet())
 }
